Excluded API. Bot connects with database directly. Improved speed up to 10x faster. Finished My questions button.

+++ data/AnswerModel.py
@@ -14,5 +14,5 @@ class Answer(SqlAlchemyBase, SerializerMixin):
     text = sqlalchemy.Column(sqlalchemy.String, unique=True, nullable=False)
     from_user_id = sqlalchemy.Column(sqlalchemy.BigInteger, sqlalchemy.ForeignKey('users.id'))
     from_user = orm.relation('User')
-    question_id = sqlalchemy.Column(sqlalchemy.BigInteger, sqlalchemy.ForeignKey('questions.id'))
+    question_id = sqlalchemy.Column(sqlalchemy.Integer, sqlalchemy.ForeignKey('questions.id'))
     question = orm.relation('Question')
